unify types with Geth

+++ src/Nethermind/Nethermind.AuRa.Test/AuRaStepCalculatorTests.cs
@@ -83,8 +83,10 @@ namespace Nethermind.AuRa.Test
         {
             DateTimeOffset now = DateTimeOffset.FromUnixTimeSeconds(BaseOffset);
             ManualTimestamper timestamper = new(now.UtcDateTime);
-            Dictionary<long, long> stepDurations = secondsDurations.ToDictionary(
-                kvp => kvp.SecondsOffset == 0 ? 0 : kvp.SecondsOffset + now.ToUnixTimeSeconds(),
+            // Keys are ulong as required by AuRaStepCalculator; SecondsOffset==0 maps to key 0, otherwise
+            // key = SecondsOffset + now.ToUnixTimeSeconds() — both non-negative, cast is safe.
+            Dictionary<ulong, long> stepDurations = secondsDurations.ToDictionary(
+                kvp => kvp.SecondsOffset == 0 ? 0UL : (ulong)(kvp.SecondsOffset + now.ToUnixTimeSeconds()),
                 kvp => kvp.StepDuration);
             AuRaStepCalculator calculator = new(stepDurations, timestamper, LimboLogs.Instance);
             timestamper.Add(TimeSpan.FromSeconds(second));
@@ -92,6 +94,7 @@ namespace Nethermind.AuRa.Test
             Assert.That(calculator.TimeToNextStep, Is.EqualTo(TimeSpan.FromSeconds(expectedSecondsToNextStep)));
         }
 
-        private IDictionary<long, long> GetStepDurationsForSingleStep(long stepDuration) => new Dictionary<long, long>() { { 0, stepDuration } };
+        private IDictionary<ulong, long> GetStepDurationsForSingleStep(long stepDuration) =>
+            new Dictionary<ulong, long>() { { 0UL, stepDuration } };
     }
 }
